Compliance: refactor handleKernelAppend duplicate
